Refactor google.api.core to google.api_core.
